use int for the magnitude
git-svn-id: svn://svn.mplayerhq.hu/soc/dirac@855 66570020-b32e-0410-962d-e44e8f9afd43

+++ libavcodec/dirac.c
@@ -770,7 +770,7 @@ static int inline coeff_quant_offset(DiracContext *s, int idx) {
  */
 static int inline coeff_dequant(int coeff,
                                 int qoffset, int qfactor) {
-    int64_t magnitude = coeff * qfactor;
+    int magnitude = coeff * qfactor;
 
     if (! magnitude)
         return 0;
